Hillshade baselayer and import .zip shapefile
Added functionality to toggle on/off hillshade layer - note that this
will only work when the correct IP address to the virtual machine with
ArcGIS Server is entered in the top of the .js file. Also added
functionality to import .zip shapefile. Began working on toggling the
imported layer, but this isn't working yet.